UADS-53 refactor: Moved CMS into Backend folder
