+++ backend/index.ts
@@ -1,25 +1,23 @@
-import express, { json } from 'express';
-import cors from 'cors';
-import { connect } from 'mongoose';
-import { config } from 'dotenv';
-
-// Import Routers
-import router from './routes/routes';
-
-const app = express();
+import express, { json } from "express";
+import cors from "cors";
+import router from "./routes/routes";
+import { config } from "dotenv";
 config();
 
-// const databaseUrl: string = process.env.DATABASE_URL!;
-// connect(databaseUrl);
+// Sets our port to the PORT .env value or 4000 by default if .env is not configured
+const PORT = process.env.PORT ?? 4000;
 
+// Creates the express server
+const app = express();
+
+// Express middleware
 app.use(json());
 app.use(cors());
-app.use(express.static('public'));
+app.use(express.static("public"));
 
 // Routes
-app.use('/', router);
+app.use("/", router);
 
-const port = Number.parseInt(process.env.PORT || '3000');
-app.listen(port, () => {
-  console.log(`Listening on port ${port}`);
+app.listen(PORT, () => {
+	console.log(`Listening on port ${PORT}`);
 });
