--- cms/index.ts
@@ -1,19 +1,0 @@
-import express from "express";
-import cors from "cors";
-import router from "./routes/routes";
-import { config } from "dotenv";
-config();
-
-// Sets our port to the PORT .env value or 4000 by default if .env is not configured
-const PORT = process.env.PORT ?? 4000;
-
-// Creates the express server
-const app = express();
-
-// Express middleware
-app.use(cors());
-app.use(express.json());
-
-app.use("/", router);
-
-app.listen(PORT, () => console.log(`Listening on port ${PORT}`));
