--- cms/routes/routes.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import sponsorRoutes from "./sponsorRoutes";
-
-const router = Router();
-
-router.use("/cms/sponsors", sponsorRoutes);
-
-export default router;
