--- cms/routes/sponsorRoutes.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import { getSponsors } from "../controllers/SponsorController";
-
-const sponsorRoutes = Router();
-
-sponsorRoutes.get("/", getSponsors);
-
-export default sponsorRoutes;
