--- cms/tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "compilerOptions": {
-      "target": "es2020",
-      "lib": ["dom", "es2020"],
-      "module": "commonjs",
-      "outDir": "./dist",
-      "strict": true,
-      "esModuleInterop": true,
-      "skipLibCheck": true
-    },
-    "include": ["*.ts"],
-    "exclude": ["node_modules"]
-  }
-
