--- notion/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "notion",
-  "version": "1.0.0",
-  "main": "sponsor.ts",
-  "type": "module",
-  "dependencies": {
-    "@notionhq/client": "^2.2.15",
-    "@types/node": "^20.14.4",
-    "acorn": "^8.12.0",
-    "acorn-walk": "^8.3.3",
-    "arg": "^4.1.3",
-    "asynckit": "^0.4.0",
-    "axios": "^1.7.2",
-    "combined-stream": "^1.0.8",
-    "create-require": "^1.1.1",
-    "delayed-stream": "^1.0.0",
-    "diff": "^4.0.2",
-    "dotenv": "^16.4.5",
-    "follow-redirects": "^1.15.6",
-    "form-data": "^4.0.0",
-    "make-error": "^1.3.6",
-    "mime-db": "^1.52.0",
-    "mime-types": "^2.1.35",
-    "node-fetch": "^2.7.0",
-    "proxy-from-env": "^1.1.0",
-    "punycode": "^2.3.1",
-    "tr46": "^0.0.3",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.4.5",
-    "undici-types": "^5.26.5",
-    "v8-compile-cache-lib": "^3.0.1",
-    "webidl-conversions": "^3.0.1",
-    "whatwg-url": "^5.0.0",
-    "yn": "^3.1.1"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.3"
-  },
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/sponsor.js",
-    "dev": "nodemon --watch src --ext ts --exec \"yarn build && node dist/sponsor.js\""
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": ""
-}
